Reorganize mongo schema structure
