+++ src/infrastructure/schemas/product/product.document.ts
@@ -1,0 +1,8 @@
+import { Document } from "mongoose";
+
+export interface IProductDocument extends Document {
+  name: string;
+  description: string;
+  price: number;
+  category: string;
+}
