+++ src/infrastructure/schemas/user/user.document.ts
@@ -1,0 +1,8 @@
+import { Document } from "mongoose";
+
+export interface IUserDocument extends Document {
+  username: string;
+  email: string;
+  password: string;
+  isAdmin: boolean;
+}
